关注粉丝

+++ src/main/java/com/mettre/account/serviceimpl/VisitorServiceImpl.java
@@ -3,7 +3,10 @@ package com.mettre.account.serviceimpl;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.mettre.account.base.ReturnType;
 import com.mettre.account.enum_.ResultEnum;
+import com.mettre.account.exception.CustomerException;
+import com.mettre.account.jwt.SecurityContextStore;
 import com.mettre.account.mapper.VisitorMapper;
+import com.mettre.account.pojo.User;
 import com.mettre.account.pojo.Visitor;
 import com.mettre.account.service.UserService;
 import com.mettre.account.service.VisitorService;
@@ -34,9 +37,12 @@ public class VisitorServiceImpl implements VisitorService {
     @Override
     public int insert(Visitor record) {
         int type = 0;
-        if (!record.getUserId().equals(record.getVisitorsUesr())) {
-            userService.selectByPrimaryKey(record.getUserId());
-            userService.selectByPrimaryKey(record.getVisitorsUesr());
+        if (!record.getUserId().equals(record.getVisitorsUser())) {
+            User user1 = userService.selectByPrimaryKey(record.getUserId());
+            User user2 = userService.selectByPrimaryKey(record.getVisitorsUser());
+            if (user1 == null || user2 == null) {
+                throw new CustomerException(ResultEnum.VISTORERROY);
+            }
             type = visitorMapper.insert(record);
         }
         return ReturnType.ReturnType(type, ResultEnum.VISTORERROY);
@@ -72,6 +78,11 @@ public class VisitorServiceImpl implements VisitorService {
 
     @Override
     public int deleteAllVisitorFromVisitorId(Long visitorId) {
+        String userId = SecurityContextStore.getContext().getUserId();
+        Visitor visitor = selectByPrimaryKey(visitorId);
+        if (!userId.equals(visitor.getVisitorsUser())) {
+            throw new CustomerException("无权删除");
+        }
         int type = visitorMapper.deleteAllVisitorFromVisitorId(visitorId);
         return ReturnType.ReturnType(type, ResultEnum.VISTORDELETEERROY);
     }
